Remove timetable feature and restrict class actions to Admin
Deleted TimetableController and related views, removing timetable management from the application. Updated authorization in ClassController and views to restrict class creation, enrollment, and editing actions to Admin users only. Adjusted dashboard and sidebar to reflect these changes and improved role-based display logic. Refactored ClassRepository methods for student and faculty class retrieval and enrollment checks.

+++ WebSIMS/Controllers/CourseController.cs
@@ -29,7 +29,7 @@ namespace WebSIMS.Controllers
                     CourseCode = c.CourseCode,
                     CourseName = c.CourseName,
                     Description = c.Description,
-                    Credits = c.Credits,
+                    Credits = (int)c.Credits,
                     Department = c.Department,
                     CreatedAt = c.CreatedAt
                 }).ToList();
@@ -113,7 +113,7 @@ namespace WebSIMS.Controllers
                 CourseCode = course.CourseCode,
                 CourseName = course.CourseName,
                 Description = course.Description,
-                Credits = course.Credits,
+                Credits = (int)course.Credits,
                 Department = course.Department,
                 CreatedAt = course.CreatedAt
             };
@@ -136,7 +136,7 @@ namespace WebSIMS.Controllers
                 CourseCode = course.CourseCode,
                 CourseName = course.CourseName,
                 Description = course.Description,
-                Credits = course.Credits,
+                Credits = (int)course.Credits,
                 Department = course.Department
             };
